Make IN operator a Postgres Operator
There is no equivilent in Postgres for an IN operator, and
Postgres operators requires only certain special characters to
be an operator. @= is choshen because it is a combination of
the json containment operators for maps @> and <@ and the equals
operator. Therefore @= says is the value on the left contained in
an element of the value on the right.