Removed variadic from agtype->int typecasting
the postgres integer typecasts no longer use variadic
in their function signatures, because any time the function
is called with more than one argument or not agtype is an error.

Created the basics of a system to cast agtype values to postgres values
that will standardize arrays and scalar casting, when array casting
is implemented.