Refactor sqrt function
Reduced the amount of conversion between types to what is needed
and no more. if the input is numeric, the result will be numeric.
Otherwise it will return float8. Float8 is faster to compute so it
is slightly more efficient. If the user wants numeric they can typecast
to numeric before calling this function.