Remove Bulk Loader
The Bulk Load tool never worked right. It broke the
schema that postgraph needs to work correctly. The
original author was aware of this for over 2 years
but would never fix it. So now its removed.

We need tools to load data for sure, and a csv loader
is one simple potential tool, but we need to rethink
this from the ground up.