Refactor Edge Datatype
Refactored the edge datatype. Instead of storing the full
label string, instead store the graph id. Since the label id
is embeded in the graph id we have all the information we need
to retrieve the label name whenever we want.

This has two effects:

1. Improved Compression - Since the graph oid is 32 bits long and
the graph name is 32 bits to store the size of the string + the
length of the string, the stored size of a vertex will be equal to
or less than the old size.

2.
  a) General Performance Improvements - The way the label name was
     added to the vertex was slow, the new way will improve performance
     in all around way.

  b) Specific Performance Improvements - While looking up the label name
     is not very slow, because of the caching mechanism, now the label name
     will only be looked up in cases were it is necessary. ie, when the
     edge_out function is called or the labels() function is called.