Implement date_bin Function
The function date_bin “bins” the input timestamp into the
specified interval (the stride) aligned with a specified origin.

date_bin(stride, source, origin)

source is a value expression of type timestamp or timestamp with
time zone. (Values of type date are cast automatically to
timestamp.) stride is a value expression of type interval.
The return value is likewise of type timestamp or timestamp with
time zone, and it marks the beginning of the bin into which the
source is placed.

In the case of full units (1 minute, 1 hour, etc.), it gives the
same result as the analogous date_trunc call, but the difference
is that date_bin can truncate to an arbitrary interval.

The stride interval must be greater than zero and cannot contain
units of month or larger.