Implement date_trunc()
The function date_trunc is conceptually similar to the trunc
function for numbers.

date_trunc(field, source [, time_zone ])

source is a value expression of type timestamp, timestamp with
time zone, or interval. (Values of type date and time are cast
automatically to timestamp or interval, respectively.) field
selects to which precision to truncate the input value. The return
value is likewise of type timestamp, timestamp with time zone, or
interval, and it has all fields that are less significant than the
selected one set to zero (or one, for day and month).