Implement OVERLAPS Operator
(start1, end1) OVERLAPS (start2, end2)
(start1, length1) OVERLAPS (start2, length2)

This expression yields true when two time periods (defined by
their endpoints) overlap, false when they do not overlap. The
endpoints can be specified as pairs of dates, times, or time
stamps; or as a date, time, or time stamp followed by an interval.
When a pair of values is provided, either the start or the end can
be written first; OVERLAPS automatically takes the earlier value
of the pair as the start. Each time period is considered to
represent the half-open interval start <= time < end, unless start
and end are equal in which case it represents that single time
instant. This means for instance that two time periods with only
an endpoint in common do not overlap.