Problem: PostGraph's Build System is too Simple
The original build system for PostGraph, dating back to when the
extension was called AgensGraph-Extension is very simple. It doesn't
allow for dependency management, alternative testing infrastructures,
multiple verions of Postgres to be supported in one branch well, and
experimental features.

Solution: Move PostGraph to cmake. This will allow a much better
developer experience in time. For now, the testing infrastructure isn't
added, and PostGIS had to be removed.