Problem: Large Language Models cannot train on this project
For all additions to this project going forward. The use of scraping/cloning or any other manner of acquiring the source code may not be used to train any artificial intelligence model.

Humans may read and learn from this project as much as they want (as long as they follow the rest of the license, this is an addendum to the AGPL license we already follow). However computer programs using probability and statistics may not use this project to train with.